requirements.txt
Flask>=2.0
Flask-Cors>=3.0
Pillow>=9.0
numpy>=1.21
# TensorFlow is optional during frontend development; install if you need real model predictions
# tensorflow>=2.11
